+++ lib/ooxml_parser/docx_parser/docx_parser.rb
@@ -4,7 +4,7 @@ require_relative 'docx_data/document_structure'
 module OoxmlParser
   class DocxParser
     def self.parse_docx(path_to_file)
-      Parser.parse(path_to_file) do
+      Parser.parse_format(path_to_file) do
         DocumentStructure.parse
       end
     end