+++ Changelog.md
@@ -9,6 +9,7 @@
 * Add a whole lot new properties to parse in charts
 * Remove usage of Linux `file` command in favor of `ruby-filemagick` gem. Better cross-platform support
 * Use ruby method to create tmp folder instead of calling Linux methods
+* Add method `OoxmlParser::Parser.parse` to parse any OOXML format with auto recognition
 
 ### Fixes
 * Fix misplaced `dxa` and `emu` units of measurements and also fix calculation `dxa` unit